Genre search functionality added and code refactored
