+++ public/css/styles.css
@@ -2,3 +2,8 @@
 :root {
   --font-family: 'Rubik', 'Helvetica', 'Arial', sans-serif;
 }
+
+.genre-image {
+  width: 20vw;
+  height: auto
+}
