--- views/genre_search.ejs
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Pathé Gaumont Cinemas</title>
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
-    <script src="/js/jquery.js" defer></script>
-    <link rel="stylesheet" href="/css/styles.css" />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous" />
-  </head>
-  <body>
-    <header>
-      <form>
-        <select class="genre">
-            <option>--Select a genre--</option>
-            <% for(let key in genres) { %>
-            <option><%= key %></option>
-            <% } %>
-        </select>
-      </form>
-    </header>
-    <main>
-      <h1>JQuery + TMDB API</h1>
-      <div id="api-test"></div>
-    </main>
-    <script src="/js/genre-search.js"></script>
-  </body>
-</html>
